third_party/blink/web_tests/fast/filesystem/simple-temporary-expected.txt

requestFileSystem TEMPORARY test.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Successfully obtained TEMPORARY FileSystem:file__0:Temporary
PASS fileSystem.name.length > 0 is true
PASS fileSystem.root.fullPath is "/"
PASS successfullyParsed is true

TEST COMPLETE
